third_party/WebKit/LayoutTests/fast/canvas/drawImage-with-valid-image.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<script>
    window.jsTestIsAsync = true;
    description("This test checks behavior of valid arguments to Canvas::drawImage that use a valid source image.");

    function ExpectedNotEnoughArgumentsMessage(num) {
        return "\"TypeError: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': 3 arguments required, but only " + num + " present.\"";
    }
    var IndexSizeError = "IndexSizeError: Index or size was negative, or greater than the allowed value.";

    // Create image
    var myImage = new Image();
    var img_src = 'resources/apple.gif';
    myImage.src = img_src;
    myImage.onload = draw;

    var bitmap;
    var ctx = document.createElement("canvas").getContext('2d');
    function draw() {
        // No arguments should get exception
        shouldThrow("ctx.drawImage()", ExpectedNotEnoughArgumentsMessage(0));

        // image argument only should get exception
        shouldThrow("ctx.drawImage(myImage)", ExpectedNotEnoughArgumentsMessage(1));

        // image argument plus one number should get exception
        shouldThrow("ctx.drawImage(myImage, 0)", ExpectedNotEnoughArgumentsMessage(2));

        // image argument plus 2 numbers
        shouldNotThrow("ctx.drawImage(myImage, 0, 0)");

        // image argument plus 4 numbers
        shouldNotThrow("ctx.drawImage(myImage, 0, 0, 20, 20)");

        // image argument plus 8 numbers
        shouldNotThrow("ctx.drawImage(myImage, 0, 0, 20, 20, 0, 0, 20, 20)");

        // image argument plus zero size
        shouldNotThrow("ctx.drawImage(myImage, 0, 0, 0, 0)");

        // image argument plus 8 numbers, zero size
        shouldNotThrow("ctx.drawImage(myImage, 0, 0, 20, 20, 0, 0, 0, 0)");

        // imageRect does not contain sourceRect on the left side
        shouldNotThrow("ctx.drawImage(myImage, -10, 0, 52, 64, 0, 0, 20, 20)");

        // imageRect does not contain sourceRect on the right side
        shouldNotThrow("ctx.drawImage(myImage, 10, 0, 52, 64, 0, 0, 20, 20)");

        // imageRect does not contain sourceRect on top
        shouldNotThrow("ctx.drawImage(myImage, 0, -10, 52, 64, 0, 0, 20, 20)");

        // imageRect does not contain sourceRect on bottom
        shouldNotThrow("ctx.drawImage(myImage, 0, 10, 52, 64, 0, 0, 20, 20)");

        // sourceRect is bigger than imageSource on every side
        shouldNotThrow("ctx.drawImage(myImage, -10, -10, 72, 84, 0, 0, 20, 20)");

        // negative size of source, imageRect does not contain sourceRect on the left side
        shouldNotThrow("ctx.drawImage(myImage, 42, 64, -52, -64, 0, 0, 20, 20)");

        // negative size of source, imageRect does not contain sourceRect on the right side
        shouldNotThrow("ctx.drawImage(myImage, 62, 64, -52, -64, 0, 0, 20, 20)");

        // negative size of source, imageRect does not contain sourceRect on top
        shouldNotThrow("ctx.drawImage(myImage, 52, 54, -52, -64, 0, 0, 20, 20)");

        // negative size of source, imageRect does not contain sourceRect on bottom
        shouldNotThrow("ctx.drawImage(myImage, 52, 74, -52, -64, 0, 0, 20, 20)");

        // negative size of source, imageRect does not contain sourceRect on every side
        shouldNotThrow("ctx.drawImage(myImage, 62, 74, -72, -84, 0, 0, 20, 20)");

        // images with no src can be drawn
        shouldNotThrow("ctx.drawImage(new Image(), 0, 0)");
        shouldNotThrow("ctx.drawImage(new Image(), 0, 0, 20, 20)");
        shouldNotThrow("ctx.drawImage(new Image(), 0, 0, 20, 20, 0, 0, 20, 20)");

        // images with no src exit early before IndexSizeError is thrown
        shouldNotThrow("ctx.drawImage(new Image(), 0, 0)");
        shouldNotThrow("ctx.drawImage(new Image(), 0, 0, 0, 20)");
        shouldNotThrow("ctx.drawImage(new Image(), 0, 0, 0, 20, 0, 0, 20, 20)");

        createImageBitmap(myImage).then(drawBitmap, function() {
            testFailed("Promise was rejected.");
            finishJSTest();
        });
    }

    function drawBitmap(imageBitmap) {
        bitmap = imageBitmap;

        // bitmap argument plus 2 numbers
        shouldNotThrow("ctx.drawImage(bitmap, 0, 0)");

        // bitmap argument plus 4 numbers
        shouldNotThrow("ctx.drawImage(bitmap, 0, 0, 20, 20)");

        // bitmap argument plus 8 numbers
        shouldNotThrow("ctx.drawImage(bitmap, 0, 0, 20, 20, 0, 0, 20, 20)");

        // bitmap argument plus zero size
        shouldNotThrow("ctx.drawImage(bitmap, 0, 0, 0, 0)");

        // bitmap argument plus 8 numbers, zero size
        shouldNotThrow("ctx.drawImage(bitmap, 0, 0, 20, 20, 0, 0, 0, 0)");

        // bitmap argument plus 8 numbers, negative size of source, zero size
        shouldNotThrow("ctx.drawImage(bitmap, 20, 20, -20, 0, 0, 0, 20, 20)");

        // bitmap argument plus 8 numbers, negative size of destination, zero size
        shouldNotThrow("ctx.drawImage(bitmap, 0, 0, 20, 0, 20, 20, -20, -20)");

        // bitmap argument plus 8 numbers, negative size of source and destination, zero size
        shouldNotThrow("ctx.drawImage(bitmap, 20, 20, -20, 0, 20, 20, -20, -20)");

        // imageRect does not contain sourceRect on the left side
        shouldNotThrow("ctx.drawImage(bitmap, -10, 0, 52, 64, 0, 0, 20, 20)");

        // imageRect does not contain sourceRect on the right side
        shouldNotThrow("ctx.drawImage(bitmap, 10, 0, 52, 64, 0, 0, 20, 20)");

        // imageRect does not contain sourceRect on top
        shouldNotThrow("ctx.drawImage(bitmap, 0, -10, 52, 64, 0, 0, 20, 20)");

        // imageRect does not contain sourceRect on bottom
        shouldNotThrow("ctx.drawImage(bitmap, 0, 10, 52, 64, 0, 0, 20, 20)");

        // sourceRect is bigger than imageSource on every side
        shouldNotThrow("ctx.drawImage(bitmap, -10, -10, 72, 84, 0, 0, 20, 20)");

        // negative size of source, imageRect does not contain sourceRect on the left side
        shouldNotThrow("ctx.drawImage(bitmap, 42, 64, -52, -64, 0, 0, 20, 20)");

        // negative size of source, imageRect does not contain sourceRect on the right side
        shouldNotThrow("ctx.drawImage(bitmap, 62, 64, -52, -64, 0, 0, 20, 20)");

        // negative size of source, imageRect does not contain sourceRect on top
        shouldNotThrow("ctx.drawImage(bitmap, 52, 54, -52, -64, 0, 0, 20, 20)");

        // negative size of source, imageRect does not contain sourceRect on bottom
        shouldNotThrow("ctx.drawImage(bitmap, 52, 74, -52, -64, 0, 0, 20, 20)");

        // negative size of source, imageRect does not contain sourceRect on every side
        shouldNotThrow("ctx.drawImage(bitmap, 62, 74, -72, -84, 0, 0, 20, 20)");

        finishJSTest();
    }
</script>
</body>
</html>
